Enforce Proficiency Management user permissions
- also add empty table message

+++ app/Http/Controllers/ProficiencyController.php
@@ -5,11 +5,14 @@ namespace App\Http\Controllers;
 use App\Models\Proficiency;
 use Illuminate\Http\Request;
 use Illuminate\Database\QueryException;
+use Illuminate\Support\Facades\Gate;
 
 class ProficiencyController extends Controller
 {
     public function index()
     {
+        Gate::authorize('view-proficiencies');
+
         $proficiencies = Proficiency::all();
         $title = "Proficiencies";
         return view('proficiencies', compact('proficiencies', 'title'));
@@ -17,6 +20,8 @@ class ProficiencyController extends Controller
 
     public function store(Request $request)
     {
+        Gate::authorize('create-proficiencies');
+
         Proficiency::create($request->validate([
             'name' => 'required|unique:proficiencies,name|max:255',
             'description' => 'nullable|string',
@@ -27,6 +32,8 @@ class ProficiencyController extends Controller
 
     public function update(Request $request, Proficiency $proficiency)
     {
+        Gate::authorize('edit-proficiencies');
+
         $proficiency->update($request->validate([
             'name' => 'required|max:255|unique:proficiencies,name,' . $proficiency->id,
             'description' => 'nullable|string',
@@ -37,6 +44,8 @@ class ProficiencyController extends Controller
 
     public function destroy(Proficiency $proficiency)
     {
+        Gate::authorize('delete-proficiencies');
+        
         try {
             $proficiency->delete();
             return back()->with('success', 'Proficiency Deleted Successfully!');
